fix(backend): circular import for cached graphql impl
