+++ backend/infrahub/api/query.py
@@ -13,7 +13,6 @@ from infrahub.core.protocols import CoreGraphQLQuery
 from infrahub.database import InfrahubDatabase  # noqa: TC001
 from infrahub.graphql.analyzer import InfrahubGraphQLQueryAnalyzer
 from infrahub.graphql.api.dependencies import build_graphql_query_permission_checker
-from infrahub.graphql.app import graphql_impl
 from infrahub.graphql.initialization import prepare_graphql_params
 from infrahub.graphql.metrics import (
     GRAPHQL_DURATION_METRICS,
@@ -25,7 +24,7 @@ from infrahub.graphql.metrics import (
     GRAPHQL_TOP_LEVEL_QUERIES_METRICS,
 )
 from infrahub.graphql.middleware import raise_on_mutation_on_branch_needing_rebase
-from infrahub.graphql.utils import cached_parse, extract_data
+from infrahub.graphql.utils import cached_parse, extract_data, graphql_impl
 from infrahub.groups.models import RequestGraphQLQueryGroupUpdate
 from infrahub.log import get_logger
 from infrahub.workflows.catalogue import GRAPHQL_QUERY_GROUP_UPDATE
